+++ packages/terra-functional-testing/src/webpack-server/webpack-server.js
@@ -7,12 +7,13 @@ const logger = new Logger({ prefix: '[terra-functional-testing:webpack-server]' 
 
 class WebpackServer {
   constructor(options = {}) {
-    const { host, port } = options;
+    const { useHttps, host, port, gridUrl } = options;
 
+    const protocol = (useHttps === true) ? 'https' : 'http';
     this.config = WebpackServer.config(options);
     this.host = host || '0.0.0.0';
     this.port = port || '8080';
-    this.gridStatusUrl = options.gridUrl ? `http://${options.gridUrl}:80/status` : undefined;
+    this.gridStatusUrl = gridUrl ? `${protocol}://${gridUrl}:80/status` : undefined;
   }
 
   /**